pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "toronto-visual-ai-hackathon-2025"
version = "0.1.0"
description = "Visual AI project for Toronto Hackathon 2025"
readme = "README.md"
authors = ["Ads Dawson"]
classifiers = ["Programming Language :: Python :: 3.11"]
packages = [{ include = "toronto_visual_ai_hackathon", from = "src" }]

[tool.poetry.dependencies]
huggingface-hub = ">=0.26.0,<=0.29.3"
fiftyone = ">=1.4.0"
python = "==3.11.*"
numpy = ">=1.20.0"
pandas = ">=1.3.0"
matplotlib = ">=3.4.0"
pillow = ">=8.2.0"
torch = ">=1.10.0"
torchvision = ">=0.11.0"
scikit-learn = ">=1.0.0"
opencv-python = ">=4.5.0"
jupyter = ">=1.0.0"
ipykernel = ">=6.0.0"
python-dotenv = ">=0.19.0"
transformers = ">=4.50.0"
imgaug = ">=0.4.0"
requests = ">=2.32.3"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.1.0"
pytest = ">=7.0.0"
pytest-cov = ">=3.0.0"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "PT", "SIM", "ARG"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.isort]
known-first-party = ["toronto_visual_ai_hackathon"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
